requirements.txt
ultralytics>=8.1.0
opencv-python>=4.11.0.86
python-dotenv>=1.0.0
openai>=1.23.0
gradio